requirements.txt
# Python 3.10 or higher required
streamlit>=1.22.0
requests>=2.28.0
pyarrow>=12.0.0
python-dotenv>=1.0.0